Move macOS-specific code into a subdirectory
