+++ src/bsd/darwin/meson.build
@@ -1,0 +1,9 @@
+dep_tunemu = dependency('tunemu', required: opt_tunemu, static: static)
+dep_pcap = dependency('pcap', required: opt_tunemu, static: static)
+
+if dep_tunemu.found() and dep_pcap.found()
+  deps_tincd += [dep_tunemu, dep_pcap]
+  src_tincd += files('tunemu.c')
+  cdata.set('ENABLE_TUNEMU', 1)
+endif
+
