+++ src/bsd/meson.build
@@ -13,18 +13,7 @@ check_functions += [
 
 src_tincd += files('device.c')
 
-if os_name == 'openbsd'
-  subdir('openbsd')
-endif
-
-if os_name == 'darwin'
-  dep_tunemu = dependency('tunemu', required: opt_tunemu, static: static)
-  dep_pcap = dependency('pcap', required: opt_tunemu, static: static)
-
-  if dep_tunemu.found() and dep_pcap.found()
-    deps_tincd += [dep_tunemu, dep_pcap]
-    src_tincd += files('tunemu.c')
-    cdata.set('ENABLE_TUNEMU', 1)
-  endif
+if os_name in ['openbsd', 'darwin']
+  subdir(os_name)
 endif
